Docx_anon/server/requirements.txt
flask
python-multipart
aiofiles
uvicorn
jinja2
python-docx
natasha
phonenumbers